Read system keyring by default
This commit introduces a new function to read the system keyring by
default. The purpose is to have keyring defaults that allow apko to
adapt to its builder while also not requiring the user to have them
defined in the config file.

If the config file defines a set of keys, these will be used as the
image keyring, if not apko will now try to find a key set from the
building environment .

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>