Fix Tekton Task release
The workflow failed when releases were created by first creating a tag,
and the goreleaser workflow would create a Release. This caused a race
(that the Tekton Task workflow would always lose) resulting in the
task.yaml being unable to be attached to the (non-existent-at-the-time)
release.

This changes the Task generation workflow to depend on the CLI (and
release-creating) workflow, so the release is guaranteed to exist at the
time we attach to it.