Move purl value generation to SBOM options
This commit modifies the way we generate the purl data for the apko images, shifting the logic to the options object.

This has the benefit of ensuring we get the same purls in both standards by reusing the same logic.

It also fixes a bug where the purl names were formed incorrectly in the image SBOMs.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>