spdx: Move VCS URL to DownloadLocation
This commit moves the VCS URL obtained from the image config
repository to the DownloadLocation part of the main package
of the artifact being built (image or index).

This change makes the apko sboms a little better in what the
spec recommends and aligns them with some of the prevailing
practices.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>