Inline apk package (mostly) to use go-apk directly
This was largely replaced by go-apk already, but there is a lot of
vestigial indirection that doesn't really need to be there anymore.

Since we're no longer reusing build.Context for multiple things, this
drops Refesh() and moves a lot of initialization into build.New().

That initialization doesn't always make sense (e.g. when we build.New()
just to get access to configuration), so this also adds a
build.NewOptions() function that returns ImageConfiguration and Options
for setting up the platform specific builds. This is a little bit silly,
but we should be able to drop it once I've refactoring this a bit to
have separate paths for multi-platform vs single platform.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>