Remove WantSBOM and GenerateSBOM
This disables implicit SBOM generation and instead requires
build.Context callers to explicitly call a method to generate an SBOM
(which is what is already happening today).

WantSBOM is (seemingly) redundant with SBOMFormats.

GenerateSBOM got duplicated into GenerateImageSBOM.

The only place we could possibly be calling GenerateSBOM is in
build-minirootfs, but because it doesn't set SBOMFormats, it doesn't get
built.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>