Add a test to catch SBOM changes
We don't want to accidentally change SBOMs, so this will break if we do.

The wrapper image that cosign generates for SBOMs uses time.Now(), so we
have to pull out the layer and test that digest, which is unfortunate,
but it's easier than fixing cosign.

In order to make this test pass, we also needed to make
GenerateIndexSBOM deterministic, so this makes the architecture order
deterministic (blame go map iteration randomness for that).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>