fix: publish --stage-tags missing generated tags
The `--stage-tags FILE` flag is only including the tags explicitly set
as positional args to `apko publish`. Any tags generated by
`--package-version-tag/stem` are not included.

The `additionalTags` slice is empty when this code is reached. Instead
of shadowing `allTags` slice generated just before this step we can use
it. It includes a de-duped set of tags that includes both explicit tags from
the command line and any generated/discovered tags that come from
package-version and stemming.

Signed-off-by: joe miller <joeym@joeym.net>