Add an internal tarfs implementation
This is a fork of go-apk's memfs but with the ability to "write" tar
headers and their offsets. We use the memfs implementation mostly for
tracking file metadata (directories, permissions, users, and such) and
when we actually modify or create files.

For files that are sourced from APKs, we just seek to the file's offset
and read from that position without having to actually write anything to
disk. This speeds up installation tremendously when dealing with a large
number of files.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>