Use tarfs implementation for publish/build
Actually opt into the optimized tarfs path.

This cuts the amount of time we spend on I/O dramatically.

Also, add a check that if a file from an APK conflicts with a file that
apko wrote (e.g. alpine keys conflicting with alpine-keys), ignore that
conflict if their checksums match.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>