Add apko dot command
This is not 100% great because it relies on some pretty naive heuristics
to figure out the dependency graph; however, it does produce some pretty
useful graphs without touching any existing go-apk or apko code paths,
so I think it's a reasonable place to start.

Ideally, we would have ResolveWorld (or something like it) return a
graph with metadata about why things resolved the way they did, but
that's a more invasive change.

As is, this is great for understanding what packages would get pulled in
by apko and what they depend on and/or provide.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>