Write index as layout if target is a directory
This allows us to write out images in a much more efficient format for
other tools to consume.

Just writing the whole index as index.json is the simplest thing we
could do, and we end up dropping the image tag, but I think this is a
fine place to start.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>