Allow existing packages to replace installed pkg
Before this, when writing a new file, we allow a new package to install
files over any packages that it replaces, but we failed to check the
other direction.

If an already installed packages replaces a package we want to install,
we shouldn't fail if there are conflicting files. Instead, we should
just skip writing any files in the replaced package.

This way we don't rely on the installation order for replaces checks to
work as expected.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>