Make sure list of 'repositories' in the 'resolved.json.file' is complete.
Prior to this PR, there were only listed used repositories by any of the artifact for given arch.
On the other hand the list was used to run APKO in off-line environment -> and missing repos were leading to failure
(in rules_apko).

Eventually repositories and indexes will not be needed for creation of image from resolved file,
but, if we have the list in the file-format -> it should be complete.

Signed-off-by: Piotr Tabor <piotr.tabor@snowflake.com>