Change testdata to be a bit smaller
This has two packages, pretend-baselayout and replayout.

Replayout replaces pretend-baselayout and has a dependency on it.

This also makes those artifacts reproducible via:

$ ./hack/update-packages.sh

My plan is to check in some actual images built from these packages, and
the tzdata package is just a little too big to be comfy doing that.

Also drop some things from .gitignore that are very annoying.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>