Add expandapk.Split and use it
We have a lot of code that calls ExpandAPK. It's a convenient function,
but it's incredibly expensive because it does everything all at once.
This wasn't supposed to be exposed, but it escaped the lab.

What's worse is that we have things like ParsePackage being used all
over the place that also call ExpandAPK. Most things calling
ParsePackage only care about the .PKGINFO file, but we end up caching
tons of stuff to disk and hashing every file in the APK. This is very
wasteful.

Instead, this adds a new expandapk.Split function that just returns a
slice of readers in the order they're encountered. The first 1 (or 2 if
signed) readers are just buffered bytes. The last reader is whatever
remains from the input, which will be the data section.

This allows us to rewrite ParsePackage to be much faster by just
requiring the caller to supply a size. The caller is usually in a much
better position to know the size via Stat() or headers, so we don't have
to consume the entire reader in order to compute it.

This also adds ParsePackageInfo, which includes only the information
present in .PKGINFO. The existing struct returned by ParsePackage also
includes the APK size and the control section hash, which makes it
really confusing to interpret.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>