Add build.MultiArch.BuildPackageLists
In the terraform module, we call BuildPackageList for each arch:
https://github.com/chainguard-dev/terraform-provider-apko/blob/812478b999411c183e87fefb9a9ece9fe7e8f2c1/internal/provider/config_data_source.go#L239

And then use that package list to actually build the image, so we
separate the solving from the building.

This bypasses https://github.com/chainguard-dev/apko/pull/1180 which
does this multi-arch constraining through a different entrypoint.

This adds another method to the multi-arch stuff that we can opt into
for just the solve part.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>