Faster NewPkgResolver and GetRepositoryIndexes
I've been taking a look at where we're spending time on some mega-builds
and on my machine it takes ~250ms to call NewPkgResolver for each image.

This is fast in the grand scheme of things but not if we need to do it
thousands of times. The main culprit here is a regex, which sucks, but
we hacked around that by caching the results in a map. This takes that
one step further and introduces yet another (I know, sorry) global cache
of the things that need a regex. This is less terrible than the other
instances, IMO, because the cached functions are pure and take only
version strings as inputs.

Using a sync.Map over a map doesn't seem to be measureably slower on my
machine in the worst case (the first time we call NewPkgResolver), and
subsequent calls drop from ~250ms to ~90ms.

We can still do much better here by implement a real parser and maybe
even taking advantage of the interning stuff in go 1.23, but this is an
easy win without having to thing about it too hard.

The other change I'm sneaking in is to parallelize GetRepositoryIndexes
so we fetch every index in parallel. This saves us quite a bit of time
the first time we fetch these (we cache them for subsequent fetches).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>