begin a new APK client (#1218)
The idea here is to orient remote APK interactions around a **client**,
in order to give consumers the ability to handle dependency injection
properly (such as by injecting an HTTP client).

And the ultimate goal is to re-use this client whenever we need to
access remote indexes and packages. It'd be great if we can refine this
API enough where we can reuse this in all the places we need to interact
with APK repositories, to replace the many disparate implementations of
this logic we have in our tools today.

I haven't gone as far as to refactor other code in apko to use this
client, but I think that's a reasonable next step. So far this
implementation saves me from writing it again in an internal tool.

Feedback welcome. 😃 

cc: @imjasonh

---------

Signed-off-by: Dan Luhring <dluhring@chainguard.dev>