Canonicalize the architecture.
I noticed downstream that if I used `x86_64` in an apko yaml that the resulting OCI platform would be `x86_64` instead of `amd64`.

With this change, we always canonicalize the architecture before each switch statement to ensure that we are dealing with a canonical form of the architecture.

This adds a regression test for the OCI cases specifically.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>