Keep apk modtime (#1305)
We were overwriting all the file timestamps with a configurable
SOURCE_DATE_EPOCH, which was always the unix epoch. The only reason we
were doing this is that we were also setting the modtime of everything
to time.Now() elsewhere.

That's silly.

Instead, don't set anything to time.Now() and keep file modtimes when we
add them to the filesystem. If the APKs change, nothing will be
reproducible anyway, so keeping the data from the APK seems way better.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>