Work around sendfile bug (#1359)
In trying to debug why we kept hanging when we uploading to the
registry, I discovered that wrapping the os.File with something that
logged each io.Read call made the problem go away. That made me think
there was an inssue in the os.File WriteTo implementation, but I was
wrong. Instead, the issue is in an optimization in net/http that uses
sendfile if the request's Body is an os.File.

This wrapper just makes the os.File check fail so that we don't try to
use sendfile at all. This will make perf a little bit worse in some
situations, but we are not bottlenecked on reading from disk, so I'm not
worried about it.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>