fix(sbom): deduplicate SBOM packages by ID (#1366)
In https://github.com/chainguard-dev/melange/pull/1474, we updated our
SBOMs to have distinct SPDX packages for the APK itself, vs. the build
config used to produce it, vs. the upstream source(s) pulled in for the
build. This helps us produce more idiomatic and descriptive SPDX data.

But this caused a problem downstream in apko — when we aggregate APKs'
SBOMs, and specifically their SPDX packages, into a single list of SPDX
packages for the image, we end up with duplicate package IDs when
multiple APKs were build from the same build config or the same upstream
source.

This PR adds tests to catch this case, and it fixes the duplicate
package issue by dropping subsequent instances of a given package ID.

It also further improves on the existing SBOM-related tests.

Signed-off-by: Dan Luhring <dluhring@chainguard.dev>