Stop using real headers for side channels
I thought it was clever to use If-None-Match because it was pretty close
to what I wanted, but the way we're wrapping transports means that this
actually breaks if the Etag matches. Whoops!

Just change this to an invalid header so that this stops breaking, we
can fix it better later by making the cache not look like a transport.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>