Avoid race when mutating annotations
This is a quick fix to a race condition we're hitting when building
multi-arch images. The ImageConfiguration struct gets passed around by
value, but the collections (e.g. annotations) don't get copied, so if
multiple architectures attempt to set an annotation at the same time,
we'll hit a race. This creates a defensive copy to avoid that.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>