Disqualify conflicts with self for packages
Previously, we called disqualifyConflicts to disqualify any packages
that would conflict with whatever package we selected. This made it
possible to pick multiple packages with the same name, which we'd then
deduplicate and end up with whatever the first package was. That's
unexpected and produced solutions that were missing a package in a few
corner cases, mostly caused by old packages (in wolfi) providing things
accidentally that weren't withdrawn.

I've manually tested this change against our large corpus of images and
nothing failed to solve.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>