Improve error graph for failed solve
We were missing a couple of edges that would help make it clearer why
we were solving a constraint.

In particular, this DepError was problematic:

https://github.com/chainguard-dev/apko/blob/995ee5a40608591b2c4d2e7893abdffeca93f6ff/pkg/apk/apk/repo.go#L798

We were returning an error for the outer "pkg" and not including the
context of the constraint within that package being solved or the
failure of the selected package to solve its deps.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>