Make LockImageConfiguration multi-arch aware
Previously this would just return the intersection of archs with
warnings for any issues (or an error for fatal issues), but that only
partially locked the architectures because anything not in the
intersection would fail.

Instead we'll return a map of arch to config. This is a breaking change
but it's used in exactly one place and I'll fix that when I bump this.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>