Consider already selected packages during solve (#1406)
We ran into a very very slow solve. I can't tell if it was actually slow
or infinite, but it didn't seem to be convering very quickly. I realized
that we essentially recurse the entire graph for every constraint, even
for things we've already solved. This was surprising, because I thought
    we handled that already, but it seems like we didn't.
    
I've added an example "abseil-regression.yaml" that took over a second
    to solve (as of this writing) before this change and takes about 1ms
after this change. There's an open PR in wolfi that timed out because
    the abseil dep graph got even more gnarly and took several hours, so
    there was something (at least) quadratic going on here, oops.
    
I've deleted a test that was exercising a scenario where packages were
disallowed from providing themselves because there is no code path where
    that was actually possible (all invocations had "true" hard-coded).
    
I also ran this against a several thousand image configs and they all
produced the same results before and after this change, so I am fairly
    confident in it.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>