Cache DiscoverKeys calls via the shared cache
I discovered this slowdown recently when trying to do a whole bunch of
apko solves, which should be fast, but they were taking a while. I
hadn't profiled where we spend our time since we added this key
discovery mechanism, so I missed it.

This is probably incorrect for a very, very long-running process because
we'd miss advertised key updates but I don't expect that to happen over
long enough timespans to matter...

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>