Add SubFS implementation for Melange (#1560)
Relates to: https://github.com/chainguard-dev/melange/pull/1804
Relates to: https://github.com/chainguard-dev/melange/issues/1731

This PR is part 1 of 2 (the second part is the Melange counterpart).

We need a SubFS implementation in order to correctly identify packages
that are built by Melange when using the memFS filesystem.

Previously, we were doing `readlinkFS(pc.WorkspaceSubDir()` but now we
need to maintain a FullFS type throughout the build. Using `fs.Sub` on
the Melange side won't work because of this type assertion:
https://github.com/chainguard-dev/apko/blob/3e63d01bb69480954b6040f5470197aead3f81da/pkg/apk/tarball/write.go#L250

If that assertion fails, then xattrs are not preserved.

If there's an easier way to orchestrate this outside of adding an entire
SubFS implementation please let myself or @stevebeattie know!

Signed-off-by: egibs <20933572+egibs@users.noreply.github.com>