returns structured errors when origins conflict (#1578)
Generally, this is a user Config error. However, since this can happen
both during Config resolution, and building - it is hard for users using
chainguard.dev/apko as a library to flag this to the user as a user
error.

This PR converts these errors to use a type to make it easier for
library users to do so.

---------

Signed-off-by: Nghia Tran <tcnghia@gmail.com>