fix(tar): make writeTar reproducible for apko (#1595)
Right now, the apko tar artifact is not reproducible, if I use
`build-minirootfs` twice, on the same rendered yaml file (with locks) it
still does not make two identical tars

This is because of `mkdir` and symlinks evaluated when assembling the
rootfs:


![image](https://github.com/user-attachments/assets/feb1dd99-9970-41e5-a95b-fbaedcf1a508)

![image](https://github.com/user-attachments/assets/1a705611-6cfa-4e02-bd4f-e0c81ad34c91)

As you can see this affects only directories and symlinks, but the files
provided by the packages will keep their timestamp.

This PR will reset file's modinfo time, if it is in a range of the "last
minute"
this is just an heuristic to detect stuff done by this process.

The result is now a fully reproducible tar.gz build, from the same
locked yaml file:



![image](https://github.com/user-attachments/assets/dd67aa5c-9c79-4227-aa4c-8ca077b6c761)


![image](https://github.com/user-attachments/assets/f66b7947-2098-4b54-b57f-984fa3d28c06)

thanks to @jonjohnsonjr there is a prettier solution

---------

Signed-off-by: Luca Di Maio <luca.dimaio1@gmail.com>