Un-abstract the tarball package (#1609)
This is used by both melange and apko. Melange and apko have essentially
opposite requirements here, and there is less in common between the
implementations than is actually shared.

To keep the diff legible, this mostly keeps things in the same place,
but in a subsequent change I will be deleting this package and making
the implementation private.

The pieces of this that melange needs will be moved to melange.

Sibling PR: https://github.com/chainguard-dev/melange/pull/1906

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>