Implement generic layering
This uses a "simple" strategy of merging origins and replaces, then
taking the largest $budget groups of packages, merging the remainder.

This should not have any effect on apko configs that don't opt into
layering via the "layering" field, which is marked experimental:

```yaml
layering:
    strategy: "origin"
    budget: 10
```

This will split packages into up to 10 different layers (+1 for
everything that doesn't belong to a package).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>