Do not retry with jitter when discoveringKeys in offline mode.
Addresses: https://github.com/chainguard-dev/apko/issues/1330

Apko in offline mode was attempting to fetch keys... The code in case of
offline mode was doing a shortcut to check whether a file exists.
But the code was wrapped in a retryable http wrapper that introduced ~(4 x 30s) jitter
into the process.

In this PR -> in offline mode -> we turn off the retries around the fetch.
This way -> if file exists in the cache -> we use it.
If it doesn't - we accept it. And most likely we will use key provided explicitly using --extra_keys flag.