For multi-layer images, include all layers
Our image SBOMs currently embeds "the" (singular) layer digest and its
mediaType as a purl. For multi-layer images, this is no longer
reasonable. Instead, we'll include all the layers in the SBOM.

This is a bit tricky because the code in written to assume there is
always exactly one layer, but most of the tests pass exactly 0 layers,
and we want to support greater than one layer.

I've fixed up the tests and code to make the tests keep passing, but
we'll probably want to revisit what SBOMs look like when we start
shipping multiple layers (and write tests to enforce that).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>