refactor: slim down interface requirements for fetching packages (#1614)
We didn't actually need a whole implementation of `InstallablePackage`
just to fetch a package. This PR creates less needy interfaces and
inserts them when appropriate into our fetching implementation. This
also adds a convenience function for creating a minimally fetchable
package, which is useful if you don't already have such an
implementation in the context of your code.

Signed-off-by: Dan Luhring <dluhring@chainguard.dev>