Rewrite writeTar in terms of iterators (#1615)
Instead of hard-coding writeTar to only be able to write to a single
destination, this splits most of the logic out into a walkFS iterator
that yields files to the caller.

For writeTar, this change is a no-op, but I'll be trying to write to
multiple layers simultaneously in a coming change, and that will be very
messy to do deep inside this walking code, so turning it into an
iterators will make it much easier to read.

I also noticed that this package import accidentally changed from
pkg/apk/passwd to pkg/passwd. These packages are identical, so I deleted
the other one.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>