Use a lazy layerWriter for writing tar files (#1616)
Instead of writing out the entire layer in one shot, layerWriter exposes
a tar.Writer so that it's possible to construct a layerWriter, write to
it a bunch of times, then finalize the layerWriter to produce a layer.

This PR should be a no-op, but we'll take advantage of this for multiple
layers when we walk the filesystem and have multiple possible
layerWriters to choose from.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>