version.go: Ignore soname-versioned provides/depends for shlibs
Currently, packages that ship shared libraries in the archive will
have versioned provides for those shlibs.  These versions are
calculated from the soname version, e.g.:

  provides = so:libfoo.so.6=6

With the advent of
https://github.com/chainguard-dev/melange/pull/1871, the new versions
will reflect the package version.  This can introduce a package
resolution problem when the soname version is greater than the package
version, e.g.:

  libfoo-4.0-r1
  provides = so:libfoo.so.6=6

vs.

  libfoo-4.0-r2
  provides = so:libfoo.so.6=4.0-r2

It's clear that we should select libfoo-4.0-r2 as the most recent
package, but apko will currently select libfoo-4.0-r1 instead.

The solution proposed here is to tweak the versioned
depends/provides *iff* the version that comes after "=" does NOT end
with a release substring (-rN).  The tweak we do is simple: we prepend
a "0." to the version string, effectively making it sort after the
shlibs whose versions have a release substring.

This will make the case above be:

  libfoo-4.0-r1
  provides = so:libfoo.so.6=0.6

vs.

  libfoo-4.0-r2
  provides = so:libfoo.so.6=4.0-r2

which will force apko to select the latest package.

Signed-off-by: Sergio Durigan Junior <sergiodj@chainguard.dev>
Suggested-by: dann frazier <dann.frazier@chainguard.dev>