version.go: Ignore soname-versioned provides/depends for shlibs (#1619)
Currently, packages that ship shared libraries in the archive will have
versioned provides for those shlibs. These versions are calculated from
the soname version, e.g.:

```
provides = so:libfoo.so.6=6
```

With the advent of
https://github.com/chainguard-dev/melange/pull/1871, the new versions
will reflect the package version. This can introduce a package
resolution problem when the soname version is greater than the package
version, e.g.:

```
  libfoo-4.0-r1
  provides = so:libfoo.so.6=6
```

vs.

```
  libfoo-4.0-r2
  provides = so:libfoo.so.6=4.0-r2
```

It's clear that we should select libfoo-4.0-r2 as the most recent
package, but apko will currently select libfoo-4.0-r1 instead.

The solution proposed here is to ignore the versioned depends/provides
*iff* the version that comes after "=" does NOT contain a release
substring (`-rN`). This will effectively make the case above be:

```
  libfoo-4.0-r1
  provides = so:libfoo.so.6
```

vs.

```
  libfoo-4.0-r2
  provides = so:libfoo.so.6=4.0-r2
```

which will force apko to select the latest package.

Thanks to @dannf, @justinvreeland and @jonjohnsonjr for the brainstorm
sessions.

Signed-off-by: Sergio Durigan Junior <sergiodj@chainguard.dev>