Update default PATH to support sbin-merge (#1620)
In Wolfi, usr-merge and sbin-merge has been completed for binaries.

Whilst the order of directories in PATH does not matter, some software
is sensitive to be called from their preferred real location. Thus
update default PATH to order "/usr/bin" ahead of the symlinked
locations "/usr/sbin", "/sbin", and "/bin".

Keep all of the symlinked locations intact, to continue to allow
building old (locked) apko configurations which did not specify PATH,
which will install pre-usrmerge wolfi-baselayout. Also this should
continue to allow building other distributions which currently are not
working on implementing sbin-merge. Most notably this is to continue
compatibility with Alpine Linux.

Whilst untidy, this is the best compromise to ensure all containers
builds continue to work.

Note that PATH variable can be missing from the historical apko
configs. Meaning with this change old locked configs will continue to
be successfully built, and work, but will generate a different hash,
due to change of the default PATH ordering.

Given the 4 distribution paths can now often be just the one location,
ensure that `/usr/local/sbin:/usr/local/bin` are consistently used in
all examples and documentation. As these locations must be present to
enable one adding override binaries, and shell script wrappers.

This change is similar to the change already done in melange to fix
this subtle incompatibility:
- https://github.com/chainguard-dev/melange/pull/1907