CacheDir support for `apko lock` (#1612)
Speedup and code alignment of 'apko-lock'. 

Before this change 'apko lock' was downloading all the packages every
time and not utilising cacheDir.
After this change the cache is working. Also the fix should speedup apko
resolution of packages for regular `apko build` (not lockaed) as
now `ResolveWorld` and Installation are sharing downloaded packages (and
code).

Fixes: https://github.com/chainguard-dev/apko/issues/1475