layering: Synthesize directory timestamps
The way we currently implement layering, all the file metadata is
identical to what it would be if we installed every package into a
single layer (the old way). This is great for most things, but we've run
into one annoying situation around timestamps.

If multiple packages contain the same directory, we have to pick a
"winner" for which package's timestamp ends up in the filesystem.
Currently, that timestamp gets propagated back out to individual layers,
which annoyingly invalidates a lot of potential deduplication we'd
otherwise see.

This is a bit of a hack, but for package-ful layers, we can just
overwrite directory timestamps with the first file in that directory to
avoid these inter-layer influences on timestamps.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>