generate /etc/ld.so.cache (#1629)
`/etc/ld.so.cache` is a cache file of all of the public shared libraries
on the system. At runtime, the linker looks for libraries here first
before falling back to searching a built-in list of directories (`/lib`,
`/usr/lib`, `/usr/local/lib`, etc - see `/lib/ld-linux-x86_64.so.2
--help` or `/lib/ld-linux-aarch64.so.1 --help` for the full list).

Libraries that are neither found in this cache nor in one of those
built-in system paths will not be found by default. You can use the
`LD_LIBRARY_PATH` variable to extend the search path - but that's extra
config that needs to be set in multiple places in packages and images.
You can add more library directories to `/etc/ld.so.conf`, but that's
only used when generating the cache. If you don't generate the cache,
that config is useless.

`/etc/ld.so.cache` is normally generated by `ldconfig`, which most
package managers call after updating libraries. `apk` does this via
trigger. `apko` runs outside of the image context so it can not easily
do this.

This PR vendors in a copy of my branch of
https://github.com/dannf/ldso-cache (which is a fork of
https://github.com/chainguard-dev/ldso-cache) with support for parsing
`/etc/ld.so.conf` and generating `/etc/ld.so.cache` inside of an image.
It's not yet at feature parity with `ldconfig` - it only supports
`x86_64` and `aarch64`, and doesn't yet support hwcaps directories.

Once the interface seems stable, I'd like to move it back to its own
project.

---------

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>